Crudely test libsodium failures
Existing mocking libraries in Rust either don't allow for mocking
arbitrary functions, or they require nightly. The former is an absolute
requirement for our use-case, the latter would require us to make sure
we always have a nightly compiler available in CI which isn't always
easy. Further, it's pretty easy to argue that coverage testing should
apply to stable builds as well, so neither of these requirements seems
flexible.

To subvert this, we implement a crude-but-functional approach to testing
these failable functions: have a thread-local failure flag, set it on
tests that we want to fail on, and check it in these functions. Thanks
to `#[cfg]`, we don't have to pay a penalty for this in release mode.